pulled from dev; fixed components issues
